Made changes based off feedback
