+++ page-objects/PageManager.ts
@@ -6,6 +6,9 @@ import {PetDetailsPage} from "./petDetailsPage";
 import {PetTypePage} from './petTypesPage';
 import {SpecialtiesPage} from "./specialtiesPage";
 import {VeterinariansPage} from "./veterinariansPage";
+import { EditVeterinariansPage } from './editVetenarianPage';
+import { EditPetTypePage } from './editPetType';
+import { EditSpecialtiesPage } from './editSpeciality';
 
 export class PageManager
 {
@@ -15,8 +18,11 @@ export class PageManager
     private readonly ownersPage : OwnersPage
     private readonly petDetailsPage : PetDetailsPage
     private readonly petTypePage : PetTypePage
+    private readonly editPetTypePage : EditPetTypePage
     private readonly specialtiesPage : SpecialtiesPage
+    private readonly editSpecialtiesPage : EditSpecialtiesPage
     private readonly veterinariansPage : VeterinariansPage
+    private readonly editveterinariansPage : EditVeterinariansPage
     
 
     constructor(page: Page)
@@ -27,8 +33,11 @@ export class PageManager
         this.ownersPage = new OwnersPage(this.page)
         this.petDetailsPage = new PetDetailsPage(this.page)
         this.petTypePage = new PetTypePage(this.page)
+        this.editPetTypePage = new EditPetTypePage(page)
         this.specialtiesPage = new SpecialtiesPage(this.page)
+        this.editSpecialtiesPage = new EditSpecialtiesPage(this.page)
         this.veterinariansPage = new VeterinariansPage(this.page)
+        this.editveterinariansPage = new EditVeterinariansPage(this.page)
     }
 
     navigateTo()
@@ -51,6 +60,11 @@ export class PageManager
         return this.petTypePage
     }
 
+    onEditPetTypePage()
+    {
+        return this.editPetTypePage
+    }
+
     onPetDetailsPage()
     {
         return this.petDetailsPage
@@ -61,8 +75,18 @@ export class PageManager
         return this.veterinariansPage
     }
 
+    onEditVeterinariansPage()
+    {
+        return this.editveterinariansPage
+    }
+
     onSpecialtiesPage()
     {
         return this.specialtiesPage
+    }
+
+    onEditSpecialityPage()
+    {
+        return this.editSpecialtiesPage
     }
 }
