+++ tests/checkBoxes.spec.ts
@@ -4,7 +4,7 @@ import { PageManager } from '../page-objects/pageManager';
 test.beforeEach( async({page}) => {
   const pm = new PageManager(page)
 
-  await pm.navigateTo().runBeforeAllTestsToLoadClinicAndVerifyHomePage()
+  await pm.navigateTo().openHomePage()
   await pm.navigateTo().veterinariansPage()
 })
 
@@ -12,36 +12,37 @@ test('Test Case 1: Validate selected specialties', async ({page}) => {
   const pm = new PageManager(page)
 
   await pm.onVeterinariansPage().clickEditVetButtonForVeterinarian('Helen Leary')
-  await pm.onVeterinariansPage().validateEditVeterinarianToHaveSpecialty('radiology')
-  await pm.onVeterinariansPage().clickVeterinarianSpecialitiesDropDownAndValidateCheckBoxStatus()
+  await pm.onEditVeterinariansPage().validateVeterinarianSpecialty('radiology')
+
+  await pm.onEditVeterinariansPage().clickVeterinarianSpecialitiesDropDownAndValidateCheckBoxStatus()
 
   //select and unselect boxes
-  await pm.onVeterinariansPage().unCheckBoxVeterinarianSpecialitiesDropDown('radiology')
-  await pm.onVeterinariansPage().checkBoxVeterinarianSpecialitiesDropDown('surgery')
+  await pm.onEditVeterinariansPage().unCheckBoxVeterinarianSpecialitiesDropDown('radiology')
+  await pm.onEditVeterinariansPage().checkBoxVeterinarianSpecialitiesDropDown('surgery')
 
-  await pm.onVeterinariansPage().validateEditVeterinarianToHaveSpecialty('surgery')
+  await pm.onEditVeterinariansPage().validateVeterinarianSpecialty('surgery')
 
-  await pm.onVeterinariansPage().checkBoxVeterinarianSpecialitiesDropDown('dentistry')
+  await pm.onEditVeterinariansPage().checkBoxVeterinarianSpecialitiesDropDown('dentistry')
  
-  await pm.onVeterinariansPage().validateEditVeterinarianToHaveSpecialty('surgery, dentistry')
+  await pm.onEditVeterinariansPage().validateVeterinarianSpecialty('surgery, dentistry')
 });
 
 test('Test Case 2: Select all specialties', async ({page}) => {
   const pm = new PageManager(page)
 
   await pm.onVeterinariansPage().clickEditVetButtonForVeterinarian('Rafael Ortega')
-  await pm.onVeterinariansPage().validateEditVeterinarianToHaveSpecialty('surgery')
+  await pm.onEditVeterinariansPage().validateVeterinarianSpecialty('surgery')
 
-  await pm.onVeterinariansPage().selectDropDownAndcheckAllBoxes()
-  await pm.onVeterinariansPage().validateEditVeterinarianToHaveSpecialty('surgery, radiology, dentistry')
+  await pm.onEditVeterinariansPage().selectDropDownAndcheckAllBoxes()
+  await pm.onEditVeterinariansPage().validateVeterinarianSpecialty('surgery, radiology, dentistry')
 });
 
 test('Test Case 3: Unselect all specialties', async ({page}) => {
   const pm = new PageManager(page)
   
   await pm.onVeterinariansPage().clickEditVetButtonForVeterinarian('Linda Douglas')
+  
+  await pm.onEditVeterinariansPage().validateVeterinarianSpecialty('dentistry, surgery')
 
-  await pm.onVeterinariansPage().validateEditVeterinarianToHaveSpecialty('dentistry, surgery')
-
-  await pm.onVeterinariansPage().selectDropDownAndUncheckAllBoxesAndValidateEmpty()
+  await pm.onEditVeterinariansPage().selectDropDownAndUncheckAllBoxesAndValidateEmpty()
 });
