+++ tests/dateSelector.spec.ts
@@ -4,7 +4,7 @@ import { PageManager } from "../page-objects/pageManager";
 
 test.beforeEach(async ({ page }) => {
   const pm = new PageManager(page)
-  await pm.navigateTo().runBeforeAllTestsToLoadClinicAndVerifyHomePage()
+  await pm.navigateTo().openHomePage()
   await pm.navigateTo().ownersPage()
 });
 
@@ -13,25 +13,23 @@ test("Test Case 1: Select the desired date in the calendar", async ({page}) => {
   const pm = new PageManager(page)
 
   await pm.onOwnersPage().selectOwnerBasedOffName('Harold Davis')
-  await pm.onOwnersPage().ownerPageButtonSelector('Add new pet')
+  await pm.onOwnerInformationPage().selectAButtonNamed('Add new pet')
 
-  await pm.onOwnerInformationPage().addNameToInputAndValidateTextIconChange('Tom')
-  await pm.onOwnerInformationPage().pickASpecificDateAndValidateDoBField('2014', 'May', '2', '2014/05/02')
-  await pm.onOwnerInformationPage().selectAPetTypeAndSave('dog')
+  await pm.onPetDetailsPage().addNameToInputAndValidateTextIconChange('Tom')
+  await pm.onPetDetailsPage().pickASpecificDateAndValidateDoBField('2014', 'May', '2', '2014/05/02')
+  await pm.onPetDetailsPage().selectAPetTypeAndSave('dog')
 
-  await pm.onOwnersPage().validatePetHasNewInfoInOwnerPage('Tom')
+  await pm.onOwnerInformationPage().validatePetHasNewInfoInOwnerPage('Tom')
 
-  await pm.onOwnersPage().deleteAPetFromOwnerPageAndVerifyItNoLongerExists('Tom')
+  await pm.onOwnerInformationPage().deleteAPetFromOwnerPageAndVerifyItNoLongerExists('Tom')
 });
 
-test("Test Case 2: Select the dates of visits and validate dates order", async ({
-  page,
-}) => {
+test("Test Case 2: Select the dates of visits and validate dates order", async ({page}) => {
   const pm = new PageManager(page)
 
   await pm.onOwnersPage().selectOwnerBasedOffName('Jean Coleman')
 
-  await pm.onOwnerInformationPage().clickAddNewVisitorButtonForPet('Samantha')
+  await pm.onOwnerInformationPage().clickAddNewVisitButtonForPet('Samantha')
 
   await pm.onPetDetailsPage().validatePetNewVisitPetNameAndOwner('Samantha', 'Jean Coleman')
   
@@ -39,21 +37,21 @@ test("Test Case 2: Select the dates of visits and validate dates order", async (
   await pm.onPetDetailsPage().validateCurrentDatePopulatedInPetDetailsDoBField()
 
   await pm.onPetDetailsPage().addDescriptionForNewVisitPetDetail('dermatologists visit')
-  await pm.onPetDetailsPage().petDetailsPageButtonSelector('Add Visit')
+  await pm.onPetDetailsPage().selectAButtonNamed('Add Visit')
 
-  await pm.onPetDetailsPage().verifyCurrentDateNewVisitForPetName('Samantha')
+  await pm.onOwnerInformationPage().verifyCurrentDateNewVisitForPetName('Samantha')
 
-  await pm.onOwnerInformationPage().clickAddNewVisitorButtonForPet('Samantha')
+  await pm.onOwnerInformationPage().clickAddNewVisitButtonForPet('Samantha')
   
   await pm.onPetDetailsPage().selectDateFromCalenderDaysAgo(45)
   await pm.onPetDetailsPage().addDescriptionForNewVisitPetDetail('Massage Therapy')
-  await pm.onPetDetailsPage().petDetailsPageButtonSelector('Add Visit')
+  await pm.onPetDetailsPage().selectAButtonNamed('Add Visit')
 
-  await pm.onPetDetailsPage().validateMostRecentDateIsTopOfVisitListToBeTrueForPet('Samantha')
+  await pm.onOwnerInformationPage().validateMostRecentDateIsTopOfVisitListToBeTrueForPet('Samantha')
 
-  await pm.onPetDetailsPage().deleteLatestVisitorForPet('Samantha')
-  await pm.onPetDetailsPage().deleteLatestVisitorForPet('Samantha')
+  await pm.onOwnerInformationPage().deleteLatestVisitorForPet('Samantha')
+  await pm.onOwnerInformationPage().deleteLatestVisitorForPet('Samantha')
 
-  await pm.onPetDetailsPage().validateNewVisitNoLongerExistsForPetNameWithDescription('Samantha', 'dermatologists visit')
-  await pm.onPetDetailsPage().validateNewVisitNoLongerExistsForPetNameWithDescription('Samantha', 'Massage Therapy')
+  await pm.onOwnerInformationPage().validateNewVisitNoLongerExistsForPetNameWithDescription('Samantha', 'dermatologists visit')
+  await pm.onOwnerInformationPage().validateNewVisitNoLongerExistsForPetNameWithDescription('Samantha', 'Massage Therapy')
 });
