+++ tests/listsAndDropDown.spec.ts
@@ -3,7 +3,7 @@ import { PageManager } from "../page-objects/pageManager";
 
 test.beforeEach(async ({ page }) => {
   const pm = new PageManager(page)
-  await pm.navigateTo().runBeforeAllTestsToLoadClinicAndVerifyHomePage()
+  await pm.navigateTo().openHomePage()
   await pm.navigateTo().ownersPage()
 });
 
@@ -18,7 +18,7 @@ test("Test Case 1: Validate selected pet types from the list", async ({page}) =>
   await pm.onPetDetailsPage().validateOwnerNameAndFollowingPetInEditPetPage('George Franklin', 'cat')
 
   //list of options
-  await pm.onPetDetailsPage().checkThroughListOfPetsInTypeList()
+  await pm.onPetDetailsPage().validateListOfPetTypesCanBeSelectedAndAppearInTypeField()
 
 });
 
@@ -30,14 +30,13 @@ test("Test Case 2: Validate the pet type update", async ({ page }) => {
 
   await pm.onPetDetailsPage().validatePetNameAndPetTypeInEditPetPage('Rosy', 'dog')
   
-  await pm.onPetDetailsPage().selectPetTypeAndVerify('bird')
-  await pm.onPetDetailsPage().petDetailsPageButtonSelector('Update Pet')
+  await pm.onPetDetailsPage().selectPetTypeAndVerifyVisibleInTypeField('bird')
+  await pm.onPetDetailsPage().selectAButtonNamed('Update Pet')
   //not sure how to make this a reusable method
-  await expect(page.locator("app-pet-list", { hasText: "Rosy" }).locator("dd").nth(2)).toHaveText("bird");
-
+  await pm.onOwnerInformationPage().validatePetTypeForPet('Rosy', 'bird')
   await pm.onOwnerInformationPage().clickEditPetButtonForPet('Rosy')
   await pm.onPetDetailsPage().validatePetNameAndPetTypeInEditPetPage('Rosy', 'bird')
 
-  await pm.onPetDetailsPage().selectPetTypeAndVerify('dog')
-  await pm.onPetDetailsPage().petDetailsPageButtonSelector('Update Pet')
+  await pm.onPetDetailsPage().selectPetTypeAndVerifyVisibleInTypeField('dog')
+  await pm.onPetDetailsPage().selectAButtonNamed('Update Pet')
 });
